requirements.txt
MouseInfo>=0.1.3
numpy>=1.20.2
opencv-python>=4.5.1
Pillow>=8.2.0
PyAutoGUI>=0.9.52
PyGetWindow>=0.0.9
PyMsgBox>=1.0.9
pyperclip>=1.8.2
PyRect>=0.1.4
PyScreeze>=0.1.26
pytesseract>=0.3.3
PyTweening>=1.0.3
pywin32>=300
pyyaml>=6.0
requests>=2.27.1
simplejson>=3.17.6
urllib3>=1.26.9
Shapely>=1.8.2
keyboard>=0.13.5
simplejson>=3.17.6
requests>=2.28.1

